Updating collector_test.go

+++ agents/domains/evm/collector_test.go
@@ -52,5 +52,5 @@ func (i ContractSuite) TestSubmitAttestation() {
 	latestNonce, currBlock, err := attestationCollector.GetLatestNonce(i.GetTestContext(), originDomain, testDestinationDomain, i.signer)
 	Nil(i.T(), err)
 	Equal(i.T(), nonce, latestNonce)
-	Greater(i.T(), currBlock, 0)
+	Greater(i.T(), currBlock, uint64(0))
 }
